+++ examples/web_ui/frontend/src/api/types.ts
@@ -17,7 +17,6 @@ export interface TTSModelConfig {
 	type: string;
 	credential_id: string;
 	model: string;
-	realtime?: boolean | null;
 	parameters: Record<string, unknown>;
 }
