--- src/agentscope/tts/_dashscope/_cosyvoice_models/cosyvoice-v3-flash-realtime.yaml
@@ -1,24 +1,0 @@
-name: cosyvoice-v3-flash
-label: CosyVoice-v3-Flash
-status: active
-realtime: true
-
-input_types:
-  - text/plain
-
-output_types:
-  - audio/wav
-
-voices:
-  # Source: https://help.aliyun.com/en/model-studio/cosyvoice-voice-list
-  - longanyang
-  - longanhuan
-  - longhuhu_v3
-  - longyingmu_v3
-  - longxiaochun_v3
-  - longxiaoxia_v3
-  - longlaotie_v3
-  - longshuo_v3
-  - longshu_v3
-
-parameter_overrides: {}
