--- src/agentscope/tts/_dashscope/_cosyvoice_models/cosyvoice-v3-plus-realtime.yaml
@@ -1,17 +1,0 @@
-name: cosyvoice-v3-plus
-label: CosyVoice-v3-Plus
-status: active
-realtime: true
-
-input_types:
-  - text/plain
-
-output_types:
-  - audio/wav
-
-voices:
-  # Source: https://help.aliyun.com/en/model-studio/cosyvoice-voice-list
-  - longanyang
-  - longanhuan
-
-parameter_overrides: {}
